+++ programmable_video/lib/src/local_video_track.dart
@@ -92,12 +92,13 @@ class LocalVideoTrack extends VideoTrack {
   ///
   /// By default the widget will be mirrored, to change that set [mirror] to false.
   /// If you provide a [key] make sure it is unique among all [VideoTrack]s otherwise Flutter might send the wrong creation params to the native side.
-  Widget widget({bool mirror = true, Key? key}) {
+  Widget widget({bool mirror = true, VideoRenderMode mode = VideoRenderMode.BALANCED, Key? key}) {
     key ??= const ValueKey('Twilio_LocalParticipant');
 
     return ProgrammableVideoPlatform.instance.createLocalVideoTrackWidget(
       mirror: mirror,
       key: key,
+      mode: mode,
     );
   }