+++ programmable_video/lib/src/parts.dart
@@ -13,7 +13,6 @@ import 'package:twilio_programmable_video_platform_interface/twilio_programmable
 export 'package:twilio_programmable_video_platform_interface/src/audio_codecs/audio_codec.dart';
 export 'package:twilio_programmable_video_platform_interface/src/enums/enum_exports.dart';
 export 'package:twilio_programmable_video_platform_interface/src/video_codecs/video_codec.dart';
-
 export 'package:twilio_programmable_video_platform_interface/src/camera_source.dart';
 
 part 'audio_settings.dart';