+++ programmable_video/lib/src/remote_video_track.dart
@@ -25,7 +25,7 @@ class RemoteVideoTrack extends VideoTrack {
   ///
   /// By default the widget will not be mirrored, to change that set [mirror] to true.
   /// If you provide a [key] make sure it is unique among all [VideoTrack]s otherwise Flutter might send the wrong creation params to the native side.
-  Widget widget({bool mirror = false, Key? key}) {
+  Widget widget({bool mirror = false, VideoRenderMode mode = VideoRenderMode.BALANCED, Key? key}) {
     key ??= ValueKey(_sid);
     final remoteParticipantSid = _remoteParticipant.sid;
 
@@ -41,6 +41,7 @@ class RemoteVideoTrack extends VideoTrack {
       remoteParticipantSid: remoteParticipantSid,
       remoteVideoTrackSid: _sid,
       mirror: mirror,
+      mode: mode,
       key: key,
     );
   }