+++ programmable_video_platform_interface/lib/src/enums/enum_exports.dart
@@ -4,3 +4,4 @@ export 'network_quality_level.dart';
 export 'network_quality_verbosity.dart';
 export 'room_state.dart';
 export 'region.dart';
+export 'video_render_mode.dart';